Added the first version of WindNinjaLibso and a test case for trying it out, the Equilibrium Boundary Conditions (EQ-BCs) case. WindNinjaLibso contains the modified solver boundary conditions required for using log profile winds in Atmospheric Boundary Layer (ABL) flows under Equilibrium conditions, required for atmospheric flows with or without porous media in the simulation domain, required for the right k-epsilon model when no porous media is in the simulation domain. The WindNinjaLibso originally came from WindNinja, this version of WindNinjaLibso is a conversion of the original WindNinjaLibso contained in WindNinja for OpenFOAM 2.2.0, to a working version for OpenFOAM 8.
This and the next few commits to the repo will update the WindNinjaLibso and the EQ-BCs test case, showing each version of the WindNinjaLibso during the process of converting it to a working version for OpenFOAM 8. This first commit was for just the inletBCs, no additions of the ground wallBCs or updates to the k-epsilon model have yet been introduced.

It should be noted that, for this and all other cases using the WindNinjaLibso, the ending version of the WindNinjaLibso BCs should match the original WindNinja BCs pretty well, though maybe the version of the BCs could be walked back a step to use the built in ones of OpenFOAM 8 in a few cases (hence why I'm trying to keep track of the versions here with extra commits), but the numerics (fvSchemes, fvSolution) used for all these simulations are NOT the same as those used in WindNinja. The choice of numerics ended up coming from an attempt to switch from the OpenFOAM 8 windAroundBuildings tutorial numerics to the WindNinja numerics, with the end result being that a mix of the two sets of numerics were used for this EQ-BCs test case and for all the test cases in this particular porous media project. Future work could be to try to find a new way of doing the BCs and numerics to even more closely match what is done in WindNinja or to even improve what is done in WindNinja, while still improving the results of these simulations.