Updated that new k-epsilon turbulence model to now be using the formulas for all the called functions that were used in the old OpenFOAM 2.2.0 default version of k-epsilon. Had a few tricks to try to figure out which functions were getting called and when as well to make sure that they actually overloaded the inherited functions that were called before.
Note that this addition/change to the k-epsilon model caused a change to the fvSchemes used, WindNinja used a slightly different fvScheme than OpenFOAM 8 and that older scheme came back replacing the more updated version of it. This is important because going back and forth between the k-epsilon libraries can mean switching back and forth between these two fvSchemes.