Updated that new k-epsilon turbulence model, mostly just doing final cleanup. This was the final modification used for the k-epsilon model, as well as the final modification of the WindNinjaLibso and the accompanying EQ-BCs test case.
So this is the finished version of WindNinjaLibso and EQ-BCs as was used in the ember project. While it might be possible to roll back some of these updates to the WindNinjaLibso BCs and k-epsilon model, it seemed like the additions of the new nut wall model and new k-epsilon model might have been negligible, maybe less so for the new epsilon wall model but it also seemed to have little change from using the default OpenFOAM 8 versions of those additions to WindNinjaLibso with the upgrade from OpenFOAM 2.2.0 to OpenFOAM 8, but they were kept in the code as these latest versions and additions for the ember project just in case or just to be safe and consistent in the project. Future projects could briefly explore rolling back or improving the model in general.

The resulting WindNinjaLibso model was then used to know how to generate a veresion similar, but for porous media, to switch back and forth or to kind of share the porous media models with parts of the WindNinjaLibso model.