Uploaded the porous media drag forms of the WindNinja libso and simpleFoam, with accompanying verification test cases. Note that the test cases were originally done using a non-log profile version of the porous media solvers, this is the first time converting them to be log profile inlet winds instead of domain average inlet winds. So this is the first time that they have been tested with the final log profile porous media solvers. But in the process of this conversion to log profile cases, all the values were even more thoroughly evaluated to more closely match that of the papers describing the cases.
The results seem to be as expected, though the issues with the choice of numerics being oscillatory seem to be more apparent with these test cases than they were with the main cases actually used for the ember project paper. This could be a sign that the oscillatory effect was present the entire time, but masked with the slightly larger domains and larger porous media objects. Or maybe these domains are just small enough to let a bit too much inflow come in at the outflow regions on the sides and top. There could also be a slight mismatch in the Equilibrium BCs between the inlet and the ground in the first cell at the inlet, the larger gradients in that first corner cell being large enough that the convergence is more susceptible to the numerics, especially oscillatory numerics. Whatever the case, the result does appear to be consistent, just the solver takes longer to converge than might be expected as it tends to loop over the converged result rather than settling to the converged result.

This is the last and final flow solver stuff to add from the ember project, other than the cases related to the paper that also include particle simulations. So the next step is to add the particle solvers and verification test cases for the particle solver.