I'm adding the Lagrangian Particle model code, starting with the first major version, one version at a time, then adding the final version of the IBHS test case after the final version of the code up to that point. The point being to keep track of some of the old methods used and tossed out in deriving the code, the methods for the example cases isn't as important just the final version of the example cases is important.
In this version, I had already done a few different preliminary versions trying to figure out how to compile copies of the default OpenFOAM 8 code, various different organizations, and this was the first time starting to make edited versions of the code starting from those copies, one submodel of the Lagrangian Particle models at a time. For this version, only the dispersion model has been added, this version is taking the copy of the existing OpenFOAM 8 stochasticDispersionRAS module, and was editing it to try to figure out how the heck that turbulence model worked. It turned out that the default OpenFOAM 8 particle turbulence models were no good, because they were similar to the simplest version of the Langevin equations, but slightly different so not quite so physically based, at the same time they had this strange if statement with turbLoc that behaved as a somewhat random only update the velocity fluctuation once every so many timesteps. The result was that the velocity fluctuations were only turning on after particles had travelled for many many seconds, when they were getting close to the ground with higher momentum, and the velocity fluctuations were only turning on once and just holding the first value seen when they turned on. So NOT GOOD. Anyhow, this version was trying to figure out that type of stuff by commenting that if statement out and making the turbLoc and other if statement values bigger and smaller with what was originally a cone injection attempt at replicating the IBHS experimental data. It was determined that the default OpenFOAM 8 particle turbulence models were NOT adequate and that something else was needed.