Added the other submodules to the set of Lagrangian Particle code. OpenFOAM did have a normal distribution model for particle sizes, but didn't have a lognormal distribution, this was an attempt to replicate what was described online comparing to the normal distribution code. The first version of the cone injection was also added.
The myParCalcs tool needs extra description, it was designed to do runtime binning of particles into the Eulerian grid, and underwent MANY revisions, because it was insanely hard to do with the limited runtime information that was available with the current setup of the code. Honestly, if I were to have to do much more to the particle code, I would restructure the whole thing to force more of these types of information to be more accessible, which would involve major messing with the inheritance structures of the OpenFOAM Lagrangian Particle code, not just these single libraries as was done for this project. And in the end, the myParCalcs code ended up being worthless, it seemed to be correct by the end, until it was used on code with separate processors, then it DEFINITELY used the wrong algorythm to detect which particles were where and started mixing up information, after all the hard work to get the dang thing to work. So it ended up being tossed out in the end, particle binning information ended up being easier to do using post processing. But the code is still useful to get an idea of how the inheritance structure of the OpenFOAM Lagrangian particle code works, and has some weird quirky examples of how to get access to code in ways not seen in the post process functions. Technically the post process functions have less access to some of these things, the inheritance and access is such that it became harder, I just got lucky that there was an example using the post processing of particle tracking.

Anyhow, as the various versions of the myParCalcs code gets pushed to the repo, you will see just how annoying it was trying to calc the particle information with such limited runtime information. This is a similar problem with the turbulence models, but at least the amount of information normally available was exactly what was needed by the turbulence models, not so with myParCalcs.