This was the final version of the main Lagrangian Particle code.
A small update to the lognormal distribution. At this point in time it was determined that the lognormal distribution was either not correct, or the method with the desired IBHS values just couldn't get particles on the lower end of the distribution range to be injected so a switch to a normal distribution replicating the lognormal distribution was used, but the code was left just in case it could be helpful in the future, also as an example of how to make size distribution code in OpenFOAM 8.

A major update to myParCalcs. The myParCalcs seemed to finally be working, but it was determined in this last version that it looked right, until it didn't, and it seemed to be messing up when splitting the domain up into processors. So while it is still in the list of code, it was technically scrapped and calcParPositionsFromCoords was used instead to do postProcessing particle binning instead (not shown yet, that's the next push, the example code).

A major update to myConeInjection to be improved and more correct in the methods and descriptions, a more particular coneInjection to hopefully better match what was done in the IBHS test case was also added called myIBHSConeInjection, basically it was myConeInjection with slight modifications to the ways the distributions of injection angle and velocity were handled. The IBHS test case used and compared the myConeInjection version of the IBHS injection with the myIBHSConeInjection version of the IBHS injection, trying to see how the various distributions of injection angles and velocity changed the resulting ember distribution downwind.