Added the IBHS test case, used to validate the Lagrangian Particle solver against real data.
The results from this case didn't perfectly match the experimental and simulation data described in the paper describing the IBHS experiment that this was trying to replicate, but the replication attempt was good enough to know that the Lagrangian Particle solver was working correctly. In the future, the test case could be tweaked to try to get even closer to the experimental and simulation data described in the paper describing the IBHS experiment.

This case also acts as an example of how to use the Lagrangian Particle solver. It is especially useful for showing how to do angled injections and multiple injections, and in some ways uses more complicated injection methods than the final ember project cases.

To improve some of the injection methods for this case, it would require editing the injection code itself, not just the case files. See the last few commits to get examples of how to modify the code for the Lagrangian Particle model. Future work could be to try out more correct wind tunnel style BCs, especially to see if this alters the turbulence fields.