Updated the ember pan plotting script for the IBHS test case.
Just a few days ago we submitted the ember project paper to a journal, and I was required to prep an improved plot of the ember pan data, that's what this newly updated script is. The result looks a heck of a lot bigger, the plot shows details in much bigger font and it now shows the simulated data vs data described in the IBHS paper. BUT CARE SHOULD BE TAKEN WHEN USING THIS PLOT OF THE DATA DESCRIBED IN THE IBHS PAPER, the data used for this plot was an attempt to eyeball how the plot was different from the IBHS paper data plot, and so the data is NOT necessarily the same as the actual data, it might be slightly off in value especially since the plot markers were larger in the IBHS paper. The attempted replication of these data values should still be useful though.