added the first of the examples, separate from the verification test cases, plus the final update of the Lagrangian Particle solver code that was used for the project, the update just being the addition of the last few injection sources used in the ember project.
These test cases are actually a grouping of three to four different little projects, all kind of in the same realm of trying to figure out how we should do the sources for the simulations for our paper. Meaning that they are not necessarily the most ideal style of particle injections, but they could still be useful just to look at and get ideas for future simulations. See the readme for the test cases to get the information on the cases.

This set of examples technically has a lot of options and combination of options, making it less feasible to try to do all of the examples without writing some kind of automating script that also automates plotting and cleanup of simulations (particle simulations tend to fill up space quite fast). I never got around to generating those paraview plotting scripts or any automation methods, though I had some form of a start of paraview plotting scripts done for past cases that I could potentially use for a starting point, it is planned to be future work there wasn't enough time to complete such plotting scripts before wrapping up the project.