This was supposed to be added to the repo BEFORE the ember paper got published, but there was a lot going on and it has been taking a while to get this last little bit of final repo stuff for the ember paper done.
This push is just the final set of test cases, the ember paper test cases, just the examples themselves. The next step is to push the individual readme for the ember paper test cases, then to push changes to the overall readme for the repo, then finish wrapping up the ember paper test case plot scripts (I'm currently taking all the old scripts and consolidating them down into single scripts for the test cases added to the repo, but that has been taking a heck of a lot of time so may be the last thing finished, I originally wanted it to be pushed before the individual readme for the ember paper test cases), then to do one last look through to catch any last things that need cleaned up. Then the repo will be at a point where the version of it can be frozen in place as what was in the repo for the 1st ember paper publication, for easier reference, as then any changes to the repo won't be on a timeline and will just be improvements/additions for any work done on the ember stuff separate and after the 1st ember paper publication.

The only cleanup remaining SHOULD just be making sure that git didn't delete some of the empty folders that are required for OpenFOAM, so adding the keep.txt file to folders that are empty that need to be kept. Eventual work will be trying to figure out automation scripts to automate the OpenFOAM and some paraview plotting, but again that just depends on my initiative and available time whether I'll even be able to add or adjust for that stuff.